refactor: 대회 페이지 리팩토링
