--- components/organism/ApplyPage/index.ts
@@ -1,3 +1,0 @@
-import ApplyForm from './ApplyForm/ApplyForm';
-
-export { ApplyForm };
