+++ components/organism/CompetitionPage/DetailPage/ImageBanner/ImageBanner.tsx
