+++ components/organism/CompetitionPage/SuccessPage/ParticipantInfo/ParticipantInfo.tsx
