--- components/organism/SuccessPage/index.ts
@@ -1,4 +1,0 @@
-import FeeCard from './FeeCard/FeeCard';
-import ParticipantInfo from './ParticipantInfo/ParticipantInfo';
-
-export { FeeCard, ParticipantInfo };
